// tests/fixtures/types/user.type.ts
export type User = {
    username: string
    email: string
    password: string
    firstName: string
    lastName: string
    phoneNumber: string
    country: string
    city: string
    address: string
    state: string
    postalCode: string
    creditCardNumber: string
    creditCardHolderName: string
    creditCardSecurityCode: string
}